server/Code/requirements_server.txt
torch
transformers
Pillow
numpy
opencv-python
opencv-contrib-python
pygobject
pycairo
gstreamer1.0
gstreamer1.0-plugins-base
gstreamer1.0-plugins-good
gstreamer1.0-plugins-bad
gstreamer1.0-plugins-ugly
gstreamer1.0-libav
gstreamer1.0-tools
gradio
